// types/enums.ts
export enum ORDER_STATUS {
  PENDING = "PENDING",
  PROCESSING = "PROCESSING",
  FAILED = "FAILED",
  PROCESSED = "PROCESSED",
  INSUFFICIENT_CREDITS = "INSUFFICIENT_CREDITS",
}

export type PLAN = "FREE" | "MICRO" | "DISCOVERY" | "ADVENTURE" | "HEROIC";

export enum PLAN_STATUS {
  CANCELLED = "CANCELLED",
  ACTIVE = "ACTIVE",
}